+++ CHANGELOG.md
@@ -12,3 +12,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - React-flow package and workflow component. [#1](https://github.com/IN-CORE/incore-studio/issues/1)
 - Keycloak authentication to every route [#3](https://github.com/IN-CORE/incore-studio/issues/3)
 - Create home page that lists all the projects [#4](https://github.com/IN-CORE/incore-studio/issues/4)
+- Add Workflow Editor component. [#15](https://github.com/IN-CORE/incore-studio/issues/15)